refactor: Add a touch range to the hamburger menu
To make it easier on mobile, a "touch-range" was added to the
hamburger menu. This makes it easier to open the sidebar as the
user has to be a little less precise when aiming for the hamburger
menu. As long as the touch point is within the allowed range the
event to show the sidebar will trigger.