refactor: Additions/fixes specifically for mobile
- Create a fpath variable in index.js to remove duplicate code.
- Edit index.html to include markup for the percent page scrolled
mobile indicator, mobile scroll shadow, and re-done mobile topbar.
- Decrease the mobile topbar opacity to 0.85 to 0.95.
- On scroll, there is now an indicator which shows the amount the
page has been scrolled.
- On scroll, headers are tracked to be able to go back to the last
header from within the topbar.
- File content is now replaced using a replaceChild over innerHTML.
- When the alt + left/right arrow keys are used and a hash exists,
the respective header is highlighted and scrolled to.
- Before highlighting a header with a scroll animation the
highlighting class is removed.
- Due to there being a fixed topbar when an octicon anchor was
clicked on the browser would naturally scroll to the needed anchor's
element. However, since the topbar was fixed this would lead the
topbar to block the scrolled to element. To fix this the default
browser behavior is  prevented and now the scroll is done
programmatically.
- When the sidebar is scrolled to the upper/lower bounds, reset
them. Upper bound scrollTop set to 1 and lower bound scrollTop set
to (max-scroll - 1).