fix: Maintain config file order on build
Placehold file before files gets parsed/modified/worked on. Once
the promise is resolved placeholder (-1) will be replaced with
the actual worked on file contents. This is done do maintain the
file's array order. As promises end once they are resolved,
smaller files end quicker. This sometimes causes for the files
to be added in the wrong order.