refactor: Header scrolling modifications, header grouping
- Fix: Remove duplicate h5 header in cheerio query to h6.
- Feat: Wrap headers with their group contents. This causes
for a better position:sticky experience. Much like that of
iOS.
- Fix: Rename highlight animation class to
animate-header-highlight to prevent the name clash.
- Fix: Headers scrolling desktop/mobile issues.
- Fix: When attemting to header scroll, the animation is now
immediately completed when there is no change in y position.
- Style: Make the headers look better depending whether they
are in the mobile or desktop views. For example, when in the
mobile view, headers will all look the same. This is done for
a better position:sticky experience.
- Style: Make the sidebar header elements smaller and more
compact.