refactor: Font changes, Reset link-docs, Various fixes
- CSS: Make changes to font sizes and sticky header element
background/filter. Make the header space larger.
- Slow down the side menu height opening/closing animation.
- Dynamically get the hamburger menu touchstart range region
instead of having the range hard-coded.
- Add the original untouched URL to the link-doc element.
- Fix: Reset the Prism language when "md" is provided to "markup",
- Fix: On resize ensure that the overlay and sidebar hidden when
resized to a non-mobile view-port size.
- Fix: When the overlay is faded out, opacity=0, only reset the
sidebar, (add/remove class none), when in the mobile view.
- Feat: After injecting a file's contents make sure to reset the
link-docs. Basically, when the file name does not exist in the
sidebar then the file does not exist. Therefore, reset the anchor
element so that it points to the GitHub project URL.