refactor: Re-implement line highlighting to work in iOS
The new way should now also be able to work in iOS. The method used
before would not work as position: fixed was used. Using fixed
position does not play nice in iOS when inside an element that is
scrollable. The new method instead wraps the the actual line making
the need to use position: fixed not necessary.