refactor: Improve sidebar animations
We-work how the sidebar menu's elements get placed into the sidebar.
Before, all elements (menu and submenus) were all added in one go.
However, now only the main menu elements are added initially. The
submenus are then progressively added/removed on the fly to help
improve animations. As now there are significantly less elements
in the sidebar that need to be moved.