refactor: Implement missing version/missing docs error messages
- When a non-existent version is loaded, devdocs will now give some
feedback saying the version does not exist.
- To add to the previous point, the found versions will be shown.
Selecting one then load the picked version.
- When NO docs are found the message will output that no docs were
found.
- When processing the documentation and the supplied latest version
is not one of the processed versions a warning will be printed to
the console.
- The /zdata/ folder now gets purged of non processed files for that
devdocs run. Previously, all files were being left in the folder.
This was causing issues such as loading old data files. Or loading
version files that that devdocs run should not have produced.