refactor: Cleanup code block groups/placeholders, IDs, clipboard-copy
- Re-implement the way code groups are handled. More specifically,
the way the IDs are implemented. Singleton and group code blocks are
now handled in the same manner. With singleton cbs normalized to match
that of the group cbs to more easily maintain JS functionality.
- Re-implement the way the code blocks' copy-text-to-clipboard is done.
The original text is now contained via the JSON data files instead
of within the DOM via attributes to lighten up the DOM.
- Cleanup SCSS to accordingly to accommodate new changes.